+++ tests/specs/references/tupleType/inUnionType.d.ts
@@ -1,0 +1,3 @@
+export type LatLngTuple = [number, number, number?];
+
+export type LatLngExpression = string | LatLngTuple;